[FIX] CollectionView UI 에러 해결
